requirements.txt
aiohttp==3.13.1
crc32c==2.8
PyJWT==2.10.1
requests==2.32.5
pycryptodome==3.23.0
